requirements.txt
# AWS SDK
boto3==1.29.7

# Data processing
pandas==2.1.3
numpy==1.26.2

# HTTP requests
requests==2.31.0
urllib3==2.1.0

# Data formats
pyarrow==14.0.1

# Local analytics
duckdb==0.9.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2